+++ test/clientmetricreport/ClientMetricReport.test.ts
@@ -387,7 +387,7 @@ describe('ClientMetricReport', () => {
   describe('getObservableMetrics', () => {
     it('returns the observable metrics as a JS object', () => {
       const metrics = clientMetricReport.getObservableMetrics();
-      expect(Object.keys(metrics).length).to.equal(11);
+      expect(Object.keys(metrics).length).to.equal(14);
     });
   });